requirements.txt
undetected-chromedriver
selenium
natsort
pygobject # needed for pywebview
pywebview[qt]
